Combined reducers.
