+++ src/App.js
@@ -1,22 +1,22 @@
 import React from 'react';
 import './App.css';
-import {getUsers} from './actions/index.js';
+import {getUsers, getTrips} from './actions/index.js';
 import {connect} from 'react-redux';
 
 function App(props) {
   return (
     <div className="App">
       <h1>Trip Split</h1>
-      <button onClick={() => props.getUsers()}>Get Users</button>
-      {console.log("Redux users", props.users)}
+      <button onClick={() => props.getTrips()}>Get Users</button>
+      {/* {console.log("Redux trips", props)} */}
     </div>
   );
 }
 
 const mapStateToProps = (state) => {
   return {
-    users: state.users
+    trips: state.trips
   }
 }
 
-export default connect(mapStateToProps, {getUsers})(App);
+export default connect(mapStateToProps, {getTrips})(App);
