+++ src/actions/index.js
@@ -29,7 +29,7 @@ export const getTrips = () => (dispatch) => {
     dispatch({type: GET_TRIPS_START})
     axiosWithAuth().get('https://tripsplitr.herokuapp.com/trips')
             .then((res) => {
-                console.log(res.data)
+                // console.log(res.data)
                 dispatch({type: GET_TRIPS_SUCCESS, payload: res.data})
             })
             .catch((err) => dispatch({type: GET_TRIPS_FAIL, payload: err}));
