+++ src/index.js
@@ -5,10 +5,11 @@ import App from './App';
 import {BrowserRouter as Router} from 'react-router-dom';
 import {createStore, applyMiddleware} from 'redux';
 import {Provider} from 'react-redux';
-import {reducer} from './reducers/index.js';
+import rootReducer from './reducers/index.js';
 import thunk from 'redux-thunk';
+import logger from "redux-logger";
 
-const store = createStore(reducer, applyMiddleware(thunk));
+const store = createStore(rootReducer, applyMiddleware(thunk, logger));
 
 
 ReactDOM.render(
