+++ src/reducers/index.js
@@ -1,33 +1,10 @@
-import React from 'react'
-import {
-    GET_USERS_START,
-    GET_USERS_SUCCESS,
-    GET_USERS_FAIL
-} from '../actions/index.js';
+import {combineReducers} from 'redux';
+import userReducer from './userReducer.js';
+import tripReducer from './tripReducer.js';
 
-const initialState = {
-    users: [],
-    trips: []
-}
+const rootReducer = combineReducers({
+    userReducer: userReducer,
+    tripReducer: tripReducer
+})
 
-export const reducer = (state = initialState, action) => {
-    switch(action.type) {
-        case GET_USERS_START:
-            return {
-                ...state
-            }
-        
-        case GET_USERS_SUCCESS:
-            return {
-                users: action.payload
-            }
-            
-            
-        case GET_USERS_FAIL:
-            return {
-                ...state
-            }
-            
-        default: return state;
-    }
-}
+export default rootReducer;
